[FIXED JENKINS-10640] Jenkins 1.425+ takes jelly output configuration error
There was some API wreckage from a reorganization of global configuration
options[1]. This alone required the change from global.jelly -> config.jelly
but there is more.. Jenkins does not itself provide a config.jelly for the
Default Views Tab Bar, so we provide one in its namespace in order to give
relief to users of this plugin. I will attempt to commit a temporary fix
upstream so that we can remove this namespace collision.

[1] https://github.com/jenkinsci/jenkins/commit/54c0976063c8558dad74e01e364355a8c8cc24a7